requirements.txt
#fastapi
#uvicorn
#mysql-connector-python
python-dotenv
#groq
#psycopg2-binary

fastapi
uvicorn
psycopg2-binary
groq